backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9       # file uploads

# ── Database ──────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.3.2

# ── Auth ──────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
authlib==1.3.1                # Phase 7 — Google OAuth

# ── Task Queue ────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.7
flower==2.0.1

# ── AI / NLP ──────────────────────────────────────────────────────────
openai==1.35.13
spacy==3.7.5
numpy==1.26.4
sentence-transformers==3.0.1

# ── Document Parsing ──────────────────────────────────────────────────
pdfplumber==0.11.1
python-docx==1.1.2

# ── HTTP ──────────────────────────────────────────────────────────────
httpx==0.27.0
beautifulsoup4==4.12.3

# ── Config & Validation ───────────────────────────────────────────────
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

# ── Reports (Phase 7) ─────────────────────────────────────────────────
reportlab==4.2.2
openpyxl==3.1.4

# ── Utils ─────────────────────────────────────────────────────────────
python-slugify==8.0.4
aiofiles==23.2.1
structlog==24.2.0
anthropic==0.28.0
psycopg2-binary==2.9.9
rapidfuzz==3.9.3
